+++ index.html
@@ -56,6 +56,6 @@
         <div class="wall-left"></div>
         <div class="wall-right"></div>
     </div>
-    <script src="main.js"></script>
+    <script type="module" src="main.js" ></script>
 </body>
 </html>
