+++ logs.js
@@ -1,0 +1,41 @@
+const logs = {
+  start: 'Часы показывали [time], когда [player1] и [player2] бросили вызов друг другу.',
+  end: [
+    'Результат удара [playerWins]: [playerLose] - труп',
+    '[playerLose] погиб от удара бойца [playerWins]',
+    'Результат боя: [playerLose] - жертва, [playerWins] - убийца',
+  ],
+  hit: [
+    '[time], [playerDefence] пытался сконцентрироваться, но [playerKick] разбежавшись раздробил копчиком левое ухо врага. [-player.hp], [hp/100]',
+    '[time], [playerDefence] расстроился, как вдруг, неожиданно [playerKick] случайно раздробил грудью грудину противника. [-player.hp], [hp/100]',
+    '[time], [playerDefence] зажмурился, а в это время [playerKick], прослезившись, раздробил кулаком пах оппонента. [-player.hp], [hp/100]',
+    '[time], [playerDefence] чесал <вырезано цензурой>, и внезапно неустрашимый [playerKick] отчаянно размозжил грудью левый бицепс оппонента. [-player.hp], [hp/100]',
+    '[time], [playerDefence] задумался, но внезапно [playerKick] случайно влепил грубый удар копчиком в пояс оппонента. [-player.hp], [hp/100]',
+    '[time], [playerDefence] ковырялся в зубах, но [playerKick] проснувшись влепил тяжелый удар пальцем в кадык врага. [-player.hp], [hp/100]',
+    '[time], [playerDefence] вспомнил что-то важное, но внезапно [playerKick] зевнув, размозжил открытой ладонью челюсть противника. [-player.hp], [hp/100]',
+    '[time], [playerDefence] осмотрелся, и в это время [playerKick] мимоходом раздробил стопой аппендикс соперника. [-player.hp], [hp/100]',
+    '[time], [playerDefence] кашлянул, но внезапно [playerKick] показав палец, размозжил пальцем грудь соперника. [-player.hp], [hp/100]',
+    '[time], [playerDefence] пытался что-то сказать, а жестокий [playerKick] проснувшись размозжил копчиком левую ногу противника. [-player.hp], [hp/100]',
+    '[time], [playerDefence] забылся, как внезапно безумный [playerKick] со скуки, влепил удар коленом в левый бок соперника. [-player.hp], [hp/100]',
+    '[time], [playerDefence] поперхнулся, а за это [playerKick] мимоходом раздробил коленом висок врага. [-player.hp], [hp/100]',
+    '[time], [playerDefence] расстроился, а в это время наглый [playerKick] пошатнувшись размозжил копчиком губы оппонента. [-player.hp], [hp/100]',
+    '[time], [playerDefence] осмотрелся, но внезапно [playerKick] робко размозжил коленом левый глаз противника. [-player.hp], [hp/100]',
+    '[time], [playerDefence] осмотрелся, а [playerKick] вломил дробящий удар плечом, пробив блок, куда обычно не бьют оппонента. [-player.hp], [hp/100]',
+    '[time], [playerDefence] ковырялся в зубах, как вдруг, неожиданно [playerKick] отчаянно размозжил плечом мышцы пресса оппонента. [-player.hp], [hp/100]',
+    '[time], [playerDefence] пришел в себя, и в это время [playerKick] провел разбивающий удар кистью руки, пробив блок, в голень противника. [-player.hp], [hp/100]',
+    '[time], [playerDefence] пошатнулся, а в это время [playerKick] хихикая влепил грубый удар открытой ладонью по бедрам врага. [-player.hp], [hp/100]',
+  ],
+  defence: [
+    '[time], [playerKick] потерял момент и храбрый [playerDefence] отпрыгнул от удара открытой ладонью в ключицу.',
+    '[time], [playerKick] не контролировал ситуацию, и потому [playerDefence] поставил блок на удар пяткой в правую грудь.',
+    '[time], [playerKick] потерял момент и [playerDefence] поставил блок на удар коленом по селезенке.',
+    '[time], [playerKick] поскользнулся и задумчивый [playerDefence] поставил блок на тычок головой в бровь.',
+    '[time], [playerKick] старался провести удар, но непобедимый [playerDefence] ушел в сторону от удара копчиком прямо в пятку.',
+    '[time], [playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.',
+    '[time], [playerKick] не думал о бое, потому расстроенный [playerDefence] отпрыгнул от удара кулаком куда обычно не бьют.',
+    '[time], [playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.'
+  ],
+  draw: 'Ничья - это тоже победа!'
+};
+
+export default logs;
